Add removers to LearningMaterialStatus
Created LearningMaterialsEntity trait and applied it to roles and users.
