+++ src/Ilios/CoreBundle/Entity/LearningMaterialStatusInterface.php
@@ -6,6 +6,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 
 use Ilios\CoreBundle\Traits\IdentifiableEntityInterface;
+use Ilios\CoreBundle\Traits\LearningMaterialsEntityInterface;
 use Ilios\CoreBundle\Traits\TitledEntityInterface;
 
 /**
@@ -15,7 +16,8 @@ use Ilios\CoreBundle\Traits\TitledEntityInterface;
 interface LearningMaterialStatusInterface extends
     IdentifiableEntityInterface,
     TitledEntityInterface,
-    LoggableEntityInterface
+    LoggableEntityInterface,
+    LearningMaterialsEntityInterface
 {
     /**
      * @var int
@@ -31,19 +33,4 @@ interface LearningMaterialStatusInterface extends
      * @var int
      */
     const REVISED = 3;
-
-    /**
-     * @param Collection $learningMaterials
-     */
-    public function setLearningMaterials(Collection $learningMaterials);
-
-    /**
-     * @param LearningMaterialInterface $learningMaterial
-     */
-    public function addLearningMaterial(LearningMaterialInterface $learningMaterial);
-
-    /**
-     * @return ArrayCollection|LearningMaterialInterface[]
-     */
-    public function getLearningMaterials();
 }
