+++ src/Ilios/CoreBundle/Entity/LearningMaterialUserRoleInterface.php
@@ -6,6 +6,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 
 use Ilios\CoreBundle\Traits\IdentifiableEntityInterface;
+use Ilios\CoreBundle\Traits\LearningMaterialsEntityInterface;
 use Ilios\CoreBundle\Traits\TitledEntityInterface;
 
 /**
@@ -15,20 +16,7 @@ use Ilios\CoreBundle\Traits\TitledEntityInterface;
 interface LearningMaterialUserRoleInterface extends
     IdentifiableEntityInterface,
     TitledEntityInterface,
-    LoggableEntityInterface
+    LoggableEntityInterface,
+    LearningMaterialsEntityInterface
 {
-    /**
-     * @param Collection $learningMaterials
-     */
-    public function setLearningMaterials(Collection $learningMaterials);
-
-    /**
-     * @param LearningMaterialInterface $learningMaterial
-     */
-    public function addLearningMaterial(LearningMaterialInterface $learningMaterial);
-
-    /**
-     * @return ArrayCollection|LearningMaterialInterface[]
-     */
-    public function getLearningMaterials();
 }
